core(tracehouse): merge tracing-processor with trace-of-tab
