+++ lighthouse-core/computed/metrics/max-potential-fid.js
@@ -9,7 +9,7 @@ const makeComputedArtifact = require('../computed-artifact.js');
 const MetricArtifact = require('./metric.js');
 const LanternMaxPotentialFID = require('./lantern-max-potential-fid.js');
 const LHError = require('../../lib/lh-error.js');
-const TracingProcessor = require('../../lib/tracehouse/tracing-processor.js');
+const TracingProcessor = require('../../lib/tracehouse/trace-processor.js');
 
 class MaxPotentialFID extends MetricArtifact {
   /**
