+++ lighthouse-core/computed/metrics/metric.js
@@ -5,7 +5,7 @@
  */
 'use strict';
 
-const TracingProcessor = require('../../lib/tracehouse/tracing-processor.js');
+const TracingProcessor = require('../../lib/tracehouse/trace-processor.js');
 const TraceOfTab = require('../trace-of-tab.js');
 const NetworkRecords = require('../network-records.js');
