+++ lighthouse-core/test/computed/metrics/first-cpu-idle-test.js
@@ -6,7 +6,7 @@
 'use strict';
 
 const FirstCPUIdle = require('../../../computed/metrics/first-cpu-idle.js');
-const TracingProcessor = require('../../../lib/tracehouse/tracing-processor.js');
+const TracingProcessor = require('../../../lib/tracehouse/trace-processor.js');
 
 const tooShortTrace = require('../../fixtures/traces/progressive-app.json');
 const acceptableTrace = require('../../fixtures/traces/progressive-app-m60.json');
